requirements-testing.txt
python-coveralls
pytest-django
pytest-cov
pytest-pep8
django-rq
